setup.py: Yet more cleanup/simplification for the custom setup (init, mans).

+++ setup.py
@@ -2,7 +2,6 @@
 
 import os
 import subprocess
-import contextlib
 import distutils.command.clean
 
 import setuptools
@@ -12,58 +11,34 @@ import debian.changelog
 
 # Get version from debian/changelog
 MINI_BUILDD_VERSION = str(debian.changelog.Changelog(file=open("./debian/changelog", "rb")).version)
-MINI_BUILDD_MANPAGES = ["mini-buildd-tool.1", "mini-buildd.8"]
-# We need a fix static base dir for the extra files we built (to later refer to fixed static paths in debian/* debhelper files).
-MINI_BUILDD_BUILD_BASE = "build"
+MINI_BUILDD_GENFILES = {"init": "src/mini_buildd/__init__.py", "man8": "src/mini-buildd.8", "man1": "src/mini-buildd-tool.1"}
 
 
 class BuildPy(setuptools.command.build_py.build_py):
-    @classmethod
-    def _gen_init(cls):
-        path = "./src/mini_buildd/__init__.py"
-        print("I: Generating {f} ({v})...".format(f=path, v=MINI_BUILDD_VERSION))
-
-        with open(path, "w", encoding="UTF-8") as init_py:
+    def run(self):
+        print("I: Generating {}...".format(MINI_BUILDD_GENFILES["init"]))
+        with open(MINI_BUILDD_GENFILES["init"], "w", encoding="UTF-8") as init_py:
             init_py.write("""\
 # -*- coding: utf-8 -*-
 
 __version__ = "{version}"
 """.format(version=MINI_BUILDD_VERSION))
 
-    @classmethod
-    def _build_man(cls, man):
-        name, dummy, section = man.partition(".")
-        output = os.path.join(MINI_BUILDD_BUILD_BASE, man)
+        print("I: Generating {}...".format(MINI_BUILDD_GENFILES["man8"]))
+        subprocess.check_call("help2man --no-info --section 8 --include ./src/mini-buildd.help2man.include --output {} ./src/mini-buildd".format(MINI_BUILDD_GENFILES["man8"]), shell=True)
 
-        print("I: Generating \"{}\"...".format(output))
-        subprocess.check_call(["help2man",
-                               "--no-info",
-                               "--section", section,
-                               "--include", "./src/{}.help2man.include".format(name),
-                               "--output", output,
-                               "./src/{}".format(name)])
-
-    def run(self):
-        self._gen_init()
-
-        os.makedirs(MINI_BUILDD_BUILD_BASE, exist_ok=True)
-        for man in MINI_BUILDD_MANPAGES:
-            self._build_man(man)
+        print("I: Generating {}...".format(MINI_BUILDD_GENFILES["man1"]))
+        subprocess.check_call("help2man --no-info --section 1 --include ./src/mini-buildd-tool.help2man.include --output {} ./src/mini-buildd-tool".format(MINI_BUILDD_GENFILES["man1"]), shell=True)
 
         super().run()
 
 
 class Clean(distutils.command.clean.clean):
-    @classmethod
-    def _clean_man(cls, man):
-        with contextlib.suppress(FileNotFoundError):
-            f = os.path.join(MINI_BUILDD_BUILD_BASE, man)
-            print("I: Cleaning {}...".format(f))
-            os.remove(f)
-
     def run(self):
-        for man in MINI_BUILDD_MANPAGES:
-            self._clean_man(man)
+        for f in MINI_BUILDD_GENFILES.values():
+            if os.path.exists(f):
+                print("I: Cleaning {}...".format(f))
+                os.remove(f)
 
         super().run()
